requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface  
langchain-chroma
groq 
fastapi
uvicorn
pydantic
python-dotenv 
chromadb
requests
sentence-transformers 
torch
google-cloud-firestore
google-cloud-secret-manager
google-auth
beautifulsoup4
lxml
unstructured
pypdf
python-multipart